+++ src/Application/Core/IApiResponse.cs
@@ -1,0 +1,6 @@
+
+namespace ViajaNet.JobApplication.Application.Core
+{
+    public interface IApiResponse
+    { }
+}